requirements.txt
SkPy==0.7
python-telegram-bot==5.3.0